chore: update docs

+++ dev/cards/i18n.tsx
@@ -21,7 +21,7 @@ export default function ShowI18n() {
       <div>{$t('plural', { name: 'test', num: ~~(info().test % 8) })}</div>
       <br />
       <div>{$d(new Date())}</div>
-      <div>{$d(new Date(), 'short')}</div>
+      <div>{$d(new Date(), 'long')}</div>
       <div>{$d(new Date(), 'long', 'en')}</div>
       <div>{$n(100, 'currency')}</div>
     </>
